% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_function.R
\name{animate}
\alias{animate}
\title{Main function that creates animated cellular pictogram with the desired
coloring method}
\usage{
animate(
  data,
  timepoints,
  seconds,
  fps,
  dir,
  names,
  height = 250,
  width = 700,
  filename = "animation.gif",
  format
)
}
\arguments{
\item{data}{List of data structures from \link[expressyouRcell]{color_cell}}

\item{seconds}{A numerical value specifying the duration of each transition.}

\item{fps}{A numerical value specifying the number of frame in a second.}

\item{dir}{A character specifying the directory path where the datasets for
the animation are stored.}

\item{height}{gif height in pixel.}

\item{width}{gif width in pixel.}
}
\value{

}
\description{
This function creates static animated pictogram with the desired
coloring method for assigning colors to subcellular localizations
}
