% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring_methods.R
\name{assign_color_by_fdr}
\alias{assign_color_by_fdr}
\title{Plot the neuron figure coloring subcellular localizations according to FDR}
\usage{
assign_color_by_fdr(
  gene_set,
  plot_data,
  gene_loc_table,
  categorical_classes = NULL
)
}
\arguments{
\item{gene_set}{A character vector of gene names.}

\item{plot_data}{A \code{data.table} with the polygon coordinates to be
plotted.}

\item{gene_loc_table}{A \code{data.table} with information for mapping genes
to subcellular localizations.}

\item{categorical_classes}{An optional \code{data.table} with values for
specified intervals and associated colors. Default classification is
divided in 8 intervals.}
}
\value{
\itemize{A list with two \code{data.table}s:
\item{"final_dt"}{contains all the necessary information for plotting
subcellular localizations with color values depending on the FDR;}
\item{"localization_dt"}{contains a table with the enrichment scores and
pvalues obtained for each subcellular localization} }
}
\description{
This function creates the plot of the neuron figure with colored
subcellular localizations according to FDR values.
}
\examples{
#compute_enrichment(genes, plot_data, gene_loc_table)

}
