% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring_methods.R
\name{assign_color_by_value}
\alias{assign_color_by_value}
\title{Assign colors to subcellular structures according to fold changes of genes}
\usage{
assign_color_by_value(
  genes,
  plot_data,
  gene_loc_table,
  col_name,
  categorical_classes,
  coloring_mode = "mean",
  together = FALSE
)
}
\arguments{
\item{genes}{A \code{data.table} of gene names with associated log fold change values. Columns must be named "gene_symbol" and "logFC".}

\item{plot_data}{A \code{data.table} with the polygon coordinates to be plotted.}

\item{gene_loc_table}{A \code{data.table} with information for mapping genes to subcellular localizations.}

\item{coloring_mode}{Either "mean" or "median". It computes the mean (or median) of log fold change values grouping genes
by subcellular localization. Default is "mean".}

\item{colors}{A character vector of hex color codes for palette generation.}
}
\value{
A list containing a ggplot2 object and two \code{data.table}s. "final_dt" contains data associated to the
ggplot2 object, while "localization_values" contains the resulting mean (or median) fold changes for each subcellular localization.
}
\description{
This function colors subcellular structures based on mean (or median) of fold change of genes.
}
\examples{
#assign_color_by_fc(genes=genes_down,plot_data = neuron_dt,gene_loc_table = gene_loc_table,colors=c("darkred", "white"),value="median")

}
