% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{available_pictograms}
\alias{available_pictograms}
\title{Plot available cellular pictograms}
\usage{
available_pictograms()
}
\description{
This function plots the outline of the all the availables cell type.
}
\examples{
available_pictograms()

}
