% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{available_pictographs}
\alias{available_pictographs}
\title{Plot available cellular pictographs}
\usage{
available_pictographs()
}
\description{
This function plots the outline of the all the availables cell type.
}
\examples{
available_pictographs()

}
