% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{create_animation}
\alias{create_animation}
\title{Create animation between stages}
\usage{
create_animation(samples, colors, lab, n_frames, fps, output_folder = NULL)
}
\arguments{
\item{colors}{A vector of colors to be assigned to the categorical classes}

\item{lab}{A vector of labels to be assigned to the categorical classes}

\item{n_frames}{Number of frames to use in output animation}

\item{fps}{Frames per second, how many images to capture per second of video.}

\item{output_folder}{A file path where to save output animated pictures}

\item{sample_name}{A character vector with the name of the sample}

\item{plots_dt}{A list of ggplot objects with the static pictograms to be animated.}
}
\description{
This function creates animation between stages
}
\examples{
#create_animation(sample_name, colors, lab, output_folder)

}
