% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_mine_functions.R
\name{create_cell_dt}
\alias{create_cell_dt}
\title{Create neuron data.table to be plotted}
\usage{
create_cell_dt(svg_folder, colors_shapes, order_levels)
}
\arguments{
\item{svg_folder}{A character string giving the path to folder where svg files are stored}
}
\value{
the \code{data.table} with the polygon coordinates to be plotted
}
\description{
This function generates the main \code{data.table} containing all the coordinates of the polygons to be plotted
}
\examples{
#create_neuron(svg_folder="single_shapes")

}
