% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_function.R
\name{create_legend}
\alias{create_legend}
\title{Create a legend for the animated pictures}
\usage{
create_legend(color_vector, lab_vector, title)
}
\arguments{
\item{color_vector}{A vector of hex color codes.}

\item{lab_vector}{A vector of labels for the legend.}

\item{title}{A vector for the legend title.}
}
\value{
a gtable with the legend to be added on the animated pictures
}
\description{
This function creates a legend for the animated gif files.
}
