% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_mine_functions.R
\name{create_legend_dt}
\alias{create_legend_dt}
\title{Create data.table for plotting the legend of organelles}
\usage{
create_legend_dt(svg_folder)
}
\arguments{
\item{svg_folder}{A character vector with the path of the folder containing the svg files for the legend}
}
\value{
the \code{data.table} with the polygon coordinates to be plotted in the legend
}
\description{
This function creates animation between stages
}
\examples{
#create_legend_dt(svg_folder = "legend")

}
