% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{discrete_symmetric_ranges}
\alias{discrete_symmetric_ranges}
\title{Compute discrete and symmetric ranges}
\usage{
discrete_symmetric_ranges(
  timepoint_list,
  plot_data,
  gene_loc_table,
  col_name,
  grouping_vars,
  colors,
  coloring_mode,
  data_type,
  together = FALSE
)
}
\arguments{
\item{timepoint_list}{A list of \code{data.table}s, one for each time point. Each one must have at least a column
named "gene_symbol".}

\item{plot_data}{A \code{data.table} with the polygon coordinates to be plotted.}

\item{gene_loc_table}{A \code{data.table} with information for mapping genes to subcellular localizations.}

\item{col_name}{A character string with the name of the column on which the user wants to base the color of cellular
localizations when "median" or "mean" are the chosen coloring method.}

\item{grouping_vars}{A character vector with classes of genes to be considered.}

\item{colors}{A character vector with two color codes for generating a color palettes with a shade for each interval.}

\item{coloring_mode}{Either "mean" or "median". Default is "mean".

A name of a numerical columns must also be provided as \code{col_name} parameter. This method computes the mean (or
median) of values specified in the \code{col_name} column.}

\item{together}{A boolean value specifying whether genes with different classification labels should be considered
together (regardless their classification) or separately.}
}
\value{
A list of \code{data.table}s (one for each categorical variable value specified in the \code{grouping_vars}
parameter) with categorical classes into which the numerical values (previously grouped for each subcellular
localization) are mapped.
}
\description{
This function creates a table with ranges for mapping values with associated colors and labels.
}
