% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\docType{data}
\name{example_list}
\alias{example_list}
\title{Example of input time point list}
\format{
A list of \code{data.table}s with 6 columns, which are:
\describe{
\item{gene_symbol}{the x coordinate of the polygon}
\item{logFC}{the y coordinate of the polygon}
\item{sma}{index of the polygon}
\item{wt}{a string that identifies the subcellular structure}
\item{pval}{a string that uniquely identify the subpolygon}
\item{class}{a string that uniquely identify the subpolygon}
}
}
\usage{
example_list
}
\description{
An example list of two \code{data.table}s, one for each time point.
Each data.table must have at least a column of gene names named precisely “gene_symbol”.
}
\keyword{datasets}
