% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\docType{data}
\name{gene_loc_table_mm22}
\alias{gene_loc_table_mm22}
\title{Example of gene_localization table created from a mus musculus genome annotation (GENCODE v.22)}
\format{
A \code{data.table}s with 3 columns, which are:
\describe{
\item{ID}{the gene ontology term identifier}
\item{subcell_struct}{the subcellular structure name}
\item{gene_symbol}{the gene name}
}
}
\usage{
gene_loc_table_mm22
}
\description{
An example gene_localization table.
}
\keyword{datasets}
