% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R, R/legend_dt.R
\docType{data}
\name{legend_dt}
\alias{legend_dt}
\title{Legend for neuron plot data table}
\format{
A \code{data.table} with 6 columns, which are:
\describe{
\item{x}{the x coordinate of the polygon}
\item{y}{the y coordinate of the polygon}
\item{pol}{index of the polygon}
\item{subcell_struct}{a string that identifies the subcellular structure}
\item{comb}{a string that uniquely identify the subpolygon}
}

A \code{data.table} with 6 columns, which are:
\describe{
\item{x}{the x coordinate of the polygon}
\item{y}{the y coordinate of the polygon}
\item{pol}{index of the polygon}
\item{subcell_struct}{a string that identifies the subcellular structure}
\item{comb}{a string that uniquely identify the subpolygon}
}
}
\usage{
legend_dt

legend_dt
}
\description{
A \code{data.table} containing the information for plotting the legend for organelles.

A \code{data.table} containing the information for plotting the legend for organelles.
}
\keyword{datasets}
