% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_dt_nocyto.R
\docType{data}
\name{neuron_dt_nocyto}
\alias{neuron_dt_nocyto}
\title{Neuron data table without cytoplasm mapping}
\format{
A \code{data.table} with 6 columns, which are:
\describe{
\item{x}{the x coordinate of the polygon}
\item{y}{the y coordinate of the polygon}
\item{pol}{index of the polygon}
\item{organ}{a string that identifies the subcellular structure}
\item{color}{the value to fill the polygon in an example}
\item{comb}{a string that uniquely identify the subpolygon}
}
}
\usage{
neuron_dt_nocyto
}
\description{
A \code{data.table} containing the information for plotting the neuronal structure.
}
\keyword{datasets}
