% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{plot_legend_organelles}
\alias{plot_legend_organelles}
\title{Plot legend of cellular organelles}
\usage{
plot_legend_organelles(dt_legend)
}
\arguments{
\item{dt_legend}{A \code{data.table} with the polygon coordinates to be
plotted in the legend figure.}
}
\value{
A ggplot object with the legend of labeled organelles.
}
\description{
This function plots the legend of cellular organelles.
}
