% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_mine_functions.R
\name{read_coordinates}
\alias{read_coordinates}
\title{Read svg files}
\usage{
read_coordinates(file_svg)
}
\arguments{
\item{file_svg}{A character string giving the path to svg file where polygon
coordinates are saved}
}
\value{
A \code{data.table} with x and y coordinates, number of polygons and
colors
}
\description{
Read coordinates from svg file and assign them predefined
colors
}
