% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_functions.R
\name{sample_colors}
\alias{sample_colors}
\title{Sample random colors}
\usage{
sample_colors(n)
}
\arguments{
\item{n}{An integer specifying the number of random color to sample}
}
\description{
Sample random colors
}
