% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloring_methods.R
\name{assign_color_by_value}
\alias{assign_color_by_value}
\title{Assign colors to subcellular structures according to fold changes of genes}
\usage{
assign_color_by_value(
  genes,
  plot_data,
  gene_loc_table,
  col_name,
  categorical_classes,
  coloring_mode = "mean",
  together = FALSE
)
}
\arguments{
\item{genes}{A \code{data.table} of gene names with associated log fold
change values. Columns must be named "gene_symbol" and "logFC".}

\item{plot_data}{A \code{data.table} with the polygon coordinates to be
plotted.}

\item{gene_loc_table}{A \code{data.table} with information for mapping genes
to subcellular localizations.}

\item{col_name}{A character string with the name of the column on which the
user wants to base the color of cellular localizations when "median" or
"mean" are the chosen coloring method.}

\item{categorical_classes}{An optional \code{data.table} with values for
specified intervals and associated colors. Default classification is
divided in 8 intervals.}

\item{coloring_mode}{Either "mean" or "median". It computes the mean (or
median) of log fold change values grouping genes by subcellular
localization. Default is "mean".}

\item{together}{A boolean specifying whether or not genes are considered all
together without any classification. Default value is false, so genes will
be grouped by specified categories.}
}
\value{
A list containing a ggplot2 object and two \code{data.table}s.
"final_dt" contains data associated to the ggplot2 object, while
"localization_values" contains the resulting mean (or median) fold changes
for each subcellular localization.
}
\description{
This function colors subcellular structures based on mean (or
median) of fold change of genes.
}
