% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cell.R
\name{color_cell}
\alias{color_cell}
\title{Main function that creates static cellular pictogram with the desired
coloring method}
\usage{
color_cell(
  timepoint_list,
  plot_data,
  gene_loc_table,
  coloring_mode = "enrichment",
  col_name = NULL,
  colors = NULL,
  group_by = NULL,
  grouping_vars = NULL,
  ranges = NULL,
  thr = NULL,
  pval_col = NULL,
  pval_thr = NULL
)
}
\arguments{
\item{timepoint_list}{A list of \code{data.table}s, one for each time point.
Each one must have at least a column named "gene_symbol".}

\item{plot_data}{A \code{data.table} with the polygon coordinates to be
plotted.}

\item{gene_loc_table}{A \code{data.table} with information for mapping genes
to subcellular localizations.}

\item{coloring_mode}{Either "enrichment", "mean" or "median". Default is
"enrichment". If "enrichment" is specified, enrichment analysis restricted
to the sub-ontology of cellular components is performed on genes input by
the user. Colors of each subcellular compartment are based on pvalues from
the Fisher\\’s test, used to assess the statistical significance of the
enrichment.

If "mean" or "median" are specified, a name of a numerical columns must
also be provided as \code{col_name} parameter. This method computes the
mean (or median) of values specified in the \code{col_name} column.}

\item{col_name}{A character string with the name of the column on which the
user wants to base the color of cellular localizations when "median" or
"mean" are the chosen coloring method.}

\item{colors}{A character vector of hex color codes for palette generation.}

\item{group_by}{A character string with the name of the column with the
categorical variable based on which the user can compute separated analysis
on subset of genes grouped accordingly to the value of this variable.}

\item{grouping_vars}{A vector of characters for subselecting genes belonging
only to the specified values of the varible \code{group_by}}

\item{ranges}{An optional \code{data.table} with values for specified
intervals and associated colors.}
}
\value{

}
\description{
This function creates static cellular pictogram with the desired
coloring method for assigning colors to subcellular localizations
}
