% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{create_animation}
\alias{create_animation}
\title{Create animation between stages}
\usage{
create_animation(
  sample_name,
  colors,
  lab,
  plots_dt,
  n_frames,
  fps,
  output_folder = NULL
)
}
\arguments{
\item{sample_name}{A character vector with the name of the sample}

\item{colors}{A vector of colors to be assigned to the categorical classes}

\item{lab}{A vector of labels to be assigned to the categorical classes}

\item{output_folder}{A file path where to save output animated pictures}
}
\description{
This function creates animation between stages
}
\examples{
#create_animation(sample_name, colors, lab, output_folder)

}
