% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{map_gene_localization}
\alias{map_gene_localization}
\title{Create table for mapping genes to subcellular localization}
\usage{
map_gene_localization(gtf_path, dataSource = NA, organism = NA)
}
\arguments{
\item{gtf_path}{A character vector with the gtf file path.}
}
\value{
A \code{data.table} with the gene ontology subcellular localization term for each gene.
}
\description{
This function generates the \code{data.table} necessary to associate genes to a subcellular localization.
}
\details{
A gene annotatation file, in GTF format is required as input. On this complete
set of gene symbols, a gene ontology enrichment analysis is performed to associate
a gene with a term in the cellular component ontology. For this purpose, only the
sub-ontology of the cellular components is taken into consideration. This
step generates the gene-localization data.table, which maps each gene to the
locations in the cellular structures, either cellular compartments or
macromolecular complexes.
}
\examples{
#map_gene_localization(ensembl_gene_set)

}
