% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{plot_cell}
\alias{plot_cell}
\title{This function plots the outline of the selected cell type. A data.table
containing all the graphical information to draw the chosen cell type must be
specified through the coords_dt parameter. The data.tables to create various
cell types are provided within the package and the plots can be generated by
chosen the desired data structure (either cell_dt or neuron_dt). A ggplot
object with default colors for the specified cell is then given as output.}
\usage{
plot_cell(coords_dt)
}
\arguments{
\item{coords_dt}{A \code{data.table} with the polygon coordinates to be
plotted.}
}
\value{
a ggplot object with the cell figure with default colors.

A ggplot object.
}
\description{
This function creates the plot of the specified cellular
structure.
}
\examples{
plot_cell(coords_dt="neuron")

}
