% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_functions.R
\name{plot_cell}
\alias{plot_cell}
\title{Plot cellular structure with default colors}
\usage{
plot_cell(dt_complete)
}
\arguments{
\item{dt_complete}{A \code{data.table} with the polygon coordinates to be plotted.}
}
\value{
a ggplot object with the cell figure with default colors.

A ggplot object.
}
\description{
This function creates the plot of the specified cellular structure.
}
\examples{
#plot_cell(dt_complete)

}
