% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{plot_cell}
\alias{plot_cell}
\title{Plot outline of a cellular pictogram.}
\usage{
plot_cell(coords_dt)
}
\arguments{
\item{coords_dt}{A \code{data.table} with the polygon coordinates to be
plotted.}
}
\value{
a ggplot object with the cell figure with default colors.
}
\description{
This function plots the outline of the selected cell type. A \code{data.table}
containing all the graphical information to draw the chosen cell type must be
specified through the \code{coords_dt} parameter. The \code{data.tables} to create various
cell types are provided within the package and the plots can be generated by
chosen the desired data structure (\code{plot_data="cell"}, or \code{plot_data="neuron"}).
A ggplot object with default colors for the specified cell is then given as output.
}
\examples{
plot_cell(coords_dt="neuron")

}
