% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{plot_legend_organelles}
\alias{plot_legend_organelles}
\title{Plot legend of organelles in the cell}
\usage{
plot_legend_organelles(dt_legend)
}
\arguments{
\item{dt_legend}{A \code{data.table} with the polygon coordinates to be plotted in the legend figure}
}
\value{
a ggplot object with the legend of labeled organelles

A ggplot object.
}
\description{
This function plots the legend
}
\examples{
#plot_legend(dt_legend=legend)

}
